+++ App-Open-Ad/app/src/main/java/com/admob/android/ads/appopen/bestpractices/NextActivity.kt
@@ -1,0 +1,42 @@
+package com.admob.android.ads.appopen.bestpractices
+
+import android.os.Bundle
+import androidx.activity.OnBackPressedCallback
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import com.admob.android.ads.appopen.bestpractices.databinding.ActivityNextBinding
+
+class NextActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityNextBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        binding = ActivityNextBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        ViewCompat.setOnApplyWindowInsetsListener(binding.root) { v, insets ->
+            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
+            insets
+        }
+
+
+        binding.icBack.setOnClickListener {
+            onBackPressedDispatcher.onBackPressed()
+        }
+
+        onBackPressedDispatcher.addCallback(this, object : OnBackPressedCallback(true) {
+            override fun handleOnBackPressed() {
+                if (isEnabled) {
+                    isEnabled = true // Disable the callback to prevent recursive calls
+                    finish()
+                }
+            }
+        })
+    }
+
+
+}