+++ App-Open-Ad/app/src/main/java/com/admob/android/ads/appopen/bestpractices/OnShowAdCompleteListener.kt
@@ -1,0 +1,5 @@
+package com.admob.android.ads.appopen.bestpractices
+
+interface OnShowAdCompleteListener {
+    fun onShowAdComplete()
+}